+++ NEWS.md
@@ -2,6 +2,9 @@
 
 ## 3.0.1
 Released 2019-mm-dd
+
+Announcements:
+* Upgrade devel dependency `sqlite3` to version `4.0.6`
 * Log queries (https://github.com/CartoDB/CartoDB-SQL-API/pull/574)
 
 ## 3.0.0