+++ NEWS.md
@@ -7,6 +7,8 @@ Announcements:
 * Upgrade devel dependency `sqlite3` to version `4.0.6`
 * Log queries (https://github.com/CartoDB/CartoDB-SQL-API/pull/574)
 * Improve batch-queries draining while exiting the process #582
+* Implement a mechanism to short out hung connections in copy-from endpoints.
+
 
 ## 3.0.0
 Released 2019-02-22